requirements.txt
RPi.GPIO
paho-mqtt
requests